0.9

+++ button.hpp
@@ -8,11 +8,11 @@
 
 class Button {
 private:
-	int _width;
-	int _height;
+	float _width;
+	float _height;
 
-	int _posx;
-	int _posy;
+	float _posx;
+	float _posy;
 
 	bool _hover;
 	bool _click;
@@ -27,24 +27,24 @@ private:
 
 	void(*_callback)(const char& ch);
 public:
-	Button(const std::string& text = " ", const int& w = 32, const int& h = 32, const int& x = 0, const int& y = 0);
+	Button(const std::string& text = " ", const float& w = 32.0f, const float& h = 32.0f, const float& x = 0, const float& y = 0);
 	~Button();
 
-	void draw(SDL_Renderer* renderer, const int& xoffset = 0, const int& yoffset = 0) const;
+	void draw(SDL_Renderer* renderer, const float& xoffset = 0, const float& yoffset = 0) const;
 	void callback() const;
 
 	void setHover(const bool& val);
 	void setClick(const bool& val);
 	void setCallback(void(*callback)(const char& ch));
 	void setText(const std::string& text);
-	void setPosX(const int& x);
-	void setPosY(const int& y);
-	void setHeight(const int& height);
-	void setWidth(const int& width);
-	int getWidth() const;
-	int getHeight() const;
-	int getPosX() const;
-	int getPosY() const;
+	void setPosX(const float& x);
+	void setPosY(const float& y);
+	void setHeight(const float& height);
+	void setWidth(const float& width);
+	float getWidth() const;
+	float getHeight() const;
+	float getPosX() const;
+	float getPosY() const;
 	bool getClick() const;
 	std::string getText() const;
 	SDL_Color getOutlineColor() const;
